Restructure onboarding components
