+++ clients/apps/web/src/app/(main)/oauth2/authorize/components/SupportedUseCases.tsx
