+++ clients/apps/web/src/components/Onboarding/BusinessDetailsStep.tsx
@@ -21,7 +21,7 @@ import { useEffect, useMemo, useState } from 'react'
 import { useForm, useFormContext, useWatch } from 'react-hook-form'
 import slugify from 'slugify'
 import { containsBlockedWord } from '@/utils/blocked-words'
-import { CurrencySelector } from '../../CurrencySelector'
+import { CurrencySelector } from '../CurrencySelector'
 import { SUPPORTED_PAYOUT_COUNTRIES } from './config/supported-payout-countries'
 import { useOnboardingData } from './OnboardingContext'
 import { OnboardingShell } from './OnboardingShell'
