+++ clients/apps/web/src/components/Onboarding/ProductDetailsStep.tsx
