+++ clients/apps/web/src/components/Onboarding/SandboxStep.tsx
