+++ clients/apps/web/src/components/Onboarding/TermsCheckbox.tsx
